fix prop/handler name confusion
There seems to be a small confusion when explaning the above App.jsx and Note.jsx codes. The event handler function name according to the code is "toggleImportanceOf", and not "toggleImportance" (that's the name of the prop that passes the event handler to the Notes component) as is wrongly mentioned.